pyproject.toml
[project]
name = "safesf-agent"
version = "0.1.0"
description = "Multi-agent SF safety system using Claude Agent SDK and Snow Leopard"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "snowleopard>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["safesf_agent*"]
